Fix 신용 등급표 수정, mypageLayout 수정
